+++ src/controllers/PostsController.php
@@ -322,6 +322,7 @@ class PostsController extends Controller
         $fieldMappings['authorId']['attribute'] = true;
         $fieldMappings['authorId']['node'] = 'authorId/value';
         $fieldMappings['authorId']['default'] = '';
+        $fieldMappings['authorId']['options']['match'] = $uuidField;
 
         $fields['wordpressPostId']['type'] = 'text';
         $fields['wordpressPostId']['convertTo'] = 'craft\fields\PlainText';